+++ TopSecret/View/Tabs/ContentView.swift
@@ -8,7 +8,7 @@
 import SwiftUI
 
 struct ContentView: View {
-    @EnvironmentObject var viewModel : UserAuthViewModel
+    @EnvironmentObject var vm : UserAuthViewModel
     @State var selectedOption: String = "text"
     @State var isShowingNewPostView = true
     @State var selectedTab = 0
@@ -16,8 +16,8 @@ struct ContentView: View {
 
     var body: some View {
         //if there is a user signed in then go to the Tab View else go to the register view
-            if viewModel.userSession != nil {
-                NavigationView {
+            if vm.userSession != nil {
+             
                     GeometryReader { math in
                         TabView(selection: $selectedTab) {
                             VotingView()
@@ -28,7 +28,7 @@ struct ContentView: View {
                                         .frame(width: 128, height: 128)
                                 }.tag(2)
                                 .onAppear { storedTab = 2 }
-                            MessageView()
+                            MessageListView()
                                 .tabItem {
                                     Image(systemName: "message")
                                         .scaledToFill()
@@ -36,7 +36,7 @@ struct ContentView: View {
                                         .frame(width: 128, height: 128)
                                 }.tag(1)
                                 .onAppear { storedTab = 1 }
-                            HomeScreenView()
+                            HomeScreenView(vm: _vm)
                                 .tabItem {
                                     Image(systemName: "house")
 
@@ -64,32 +64,9 @@ struct ContentView: View {
                         }.accentColor(Color("TabColor"))
                         .foregroundColor(.gray)
                         .background(Color.gray)
-                        .toolbar {
-                            
-                            ToolbarItem(placement: .navigationBarLeading, content: {
-                                Button(action: { viewModel.signOut() }, label: {
-                                        Image("Hamburger_icon")
-                                        .resizable()
-                                        .frame(width: 32, height: 32)
-                                })
-                            })
-                            
-                            ToolbarItem(placement: .principal) {
-                                Image("FinishedIcon")
-                                    .resizable()
-                                    .frame(width: 64, height: 64)
-                            }
-                            
-                            ToolbarItem(placement: .navigationBarTrailing, content: {
-                                Button(action: {}, label: {
-                                    Image(systemName: "person.3.fill")
-                                        .resizable()
-                                        .frame(width: 32, height: 32)
-                                })
-                            })
-                        }
+                        
                     }
-                }.accentColor(Color("iconColor"))
+               
             }else{
                LoginView()
             }
