--- TopSecret/View/Tabs/Tabs (temp)/MessageView.swift
@@ -1,20 +1,0 @@
-//
-//  MessageView.swift
-//  TopSecret
-//
-//  Created by nathan frenzel on 8/31/21.
-//
-
-import SwiftUI
-
-struct MessageView: View {
-    var body: some View {
-        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
-    }
-}
-
-struct MessageView_Previews: PreviewProvider {
-    static var previews: some View {
-        MessageView()
-    }
-}
